Fix #4 and #5 for viewing playlists based on permissions
