+++ enchiridionapi/views/__init__.py
@@ -1,4 +1,5 @@
 from .auth import login_user, register_user
 from .episode_view import EpisodeView
 from .season_view import SeasonView
-from .playlist_view import PlaylistView
+from .user_playlist_view import UserPlaylistView
+from .public_playlist_view import PublicPlaylistView
