+++ enchiridionapi/views/user_playlist_view.py
@@ -7,21 +7,16 @@ from enchiridionapi.serializers import PlaylistSerializer, LocalEpisodeSerialize
 
 TMDB_API_KEY = os.environ.get('TMDB_API_KEY')
 
-class PlaylistView(ViewSet):
+class UserPlaylistView(ViewSet):
+
     def list(self, request):
-        """Gets a list of playlists"""
-        playlists = Playlist.objects.all()
+        if not request.user.is_authenticated:
+            return Response({"message": "You need to be logged in to view your playlists."},
+                            status=status.HTTP_403_FORBIDDEN)
+        
+        playlists = Playlist.objects.filter(user_id=request.user.id)
         serializer = PlaylistSerializer(playlists, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
-
-    def retrieve(self, request, pk=None):
-        """Handles GET requests for single playlist"""
-        try:
-            playlist = Playlist.objects.get(pk=pk)
-            serializer = PlaylistSerializer(playlist, context={'request': request})
-            return Response(serializer.data)
-        except Playlist.DoesNotExist:
-            return Response({'message': 'Playlist does not exist.'}, status=status.HTTP_404_NOT_FOUND)
 
     def create(self, request):
         user = request.user
